requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.4.2
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.2
bleach==3.1.4
blinker==1.4
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.1
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
dominate==2.5.1
exceptiongroup==1.2.2
filelock==3.16.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
ForgeryPy==0.1
frozenlist==1.5.0
git-filter-repo==2.47.0
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
humanize==4.10.0
idna==3.6
importlib_metadata==8.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jiter==0.6.1
lxml==5.3.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mechanize==0.4.8
multidict==6.1.0
names==0.3.0
openai==1.53.0
packaging==25.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
pyht==0.1.5
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==1.0.1
python-editor==1.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==1.3.5
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.1.0
virtualenv==20.29.3
visitor==0.1.3
webencodings==0.5.1
websockets==13.1
Werkzeug==0.15.5
WTForms==2.2.1
yarl==1.15.2
zipp==3.20.2
